+++ packages/snippets/ant-design-x.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/ant-design-x/index';