+++ packages/snippets/src/ant-design-x/docs/wake/welcome/demo/基本.preview.vue
@@ -1,0 +1,12 @@
+<template>
+  <Welcome
+    icon="https://mdn.alipayobjects.com/huamei_iwk9zp/afts/img/A*s5sNRo5LjfQAAAAAAAAAAAAADgCCAQ/fmt.webp"
+    title="Hello, I'm Ant Design X"
+    description="Base on Ant Design, AGI product interface solution, create a better intelligent vision~"
+  />
+</template>
+
+<script lang="ts" setup>
+  import { Welcome } from '@mixte/snippets/ant-design-x';
+  import '@mixte/snippets/ant-design-x/patch-for-react-19';
+</script>