+++ packages/snippets/src/ant-design-x/index.ts
@@ -1,0 +1,2 @@
+export * from '@mixte/snippets/ant-design-x/sender';
+export * from '@mixte/snippets/ant-design-x/welcome';