+++ packages/snippets/src/ant-design-x/info.ts
@@ -4,6 +4,9 @@ export default defineDocInfo({
   title: 'Ant Design X',
   sidebarTitle: 'Ant Design X',
   childrenGroupInfo: {
+    wake: {
+      sidebarTitle: '唤醒',
+    },
     express: {
       sidebarTitle: '表达',
     },
@@ -13,6 +16,11 @@ export default defineDocInfo({
       name: '输入框',
       title: 'Sender',
       sidebarTitle: 'Sender',
+    },
+    welcome: {
+      name: '欢迎',
+      title: 'Welcome',
+      sidebarTitle: 'Welcome',
     },
   },
 });