+++ packages/snippets/src/ant-design-x/welcome.ts
@@ -1,0 +1,9 @@
+import Welcome from './src/welcome.vue';
+
+export type WelcomeInstance = InstanceType<typeof Welcome>;
+
+export type WelcomeProps = InstanceType<typeof Welcome>['$props'];
+
+export {
+  Welcome,
+};